refactor: move catalog page's files to separate folder
