+++ src/components/view/CatalogPage/styles.scss
